Solved AoC2021 Day 23 Part 1

+++ aoc2021/aoc2021-kotlin/src/test/kotlin/de/havox_design/aoc2021/day23/Day23Test.kt
@@ -10,19 +10,19 @@ class Day23Test {
 
     @ParameterizedTest
     @MethodSource("getDataForTestProcessPart1")
-    fun testProcessPart1(filename: String, expectedResult: Long) =
+    fun testProcessPart1(filename: String, expectedResult: Int) =
         Amphipod(filename).processPart1().shouldBe(expectedResult)
 
     @ParameterizedTest
     @MethodSource("getDataForTestProcessPart2")
-    fun testProcessPart2(filename: String, expectedResult: Long) =
+    fun testProcessPart2(filename: String, expectedResult: Int) =
         Amphipod(filename).processPart2().shouldBe(expectedResult)
 
     companion object {
         @JvmStatic
         private fun getDataForTestProcessPart1(): Stream<Arguments> =
             Stream.of(
-                Arguments.of("de/havox_design/aoc2021/day23/day23part1sample.txt", 12521L)
+                Arguments.of("de/havox_design/aoc2021/day23/day23part1sample.txt", 12521)
             )
 
         @JvmStatic
